requirements.txt
# eidosSpeech v2 — Python Dependencies

# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
aiosqlite==0.20.0

# ── Settings ──────────────────────────────────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic[email]==2.10.3

# ── TTS Engine ────────────────────────────────────────────────────────────────
edge-tts==6.1.12

# ── JWT ───────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0

# ── Password Hashing ─────────────────────────────────────────────────────────
passlib[bcrypt]==1.7.4

# ── HTTP Client (Resend API, proxy health checks) ─────────────────────────────
httpx==0.28.1

# ── Email ─────────────────────────────────────────────────────────────────────
aiosmtplib==3.0.2

# ── Static files ─────────────────────────────────────────────────────────────
python-multipart==0.0.20

# ── Optional: Production ASGI ─────────────────────────────────────────────────
# gunicorn==23.0.0  # Uncomment for production
